refactor(powerpage): extract Consumption Report math to powerReport.ts
PR #166's directional integration moved calculatePowerReport's energy
+ peak figures off the net battery_power / grid_power onto the
server-derived magnitudes. Without tests that math has been
unverified by the test suite. Extract calculatePowerReport and its
helpers (PowerRow / PowerBucket / PowerReportSummary / PowerReport,
plus positivePart / integratePair / medianIntervalMs / bucketStartMs /
bucketLabel / emptyBucket / addSoc / bucketSocAvg) to src/pages/
powerReport.ts so they can be imported by vitest.

The new module has no UI/i18n dependencies: the period label is
supplied by the caller (typically powerWindowLabel(range, offset)
from PowerPage). Tests pass ''.

Behaviour is unchanged: PowerPage still owns the chart rendering,
the data fetch, and the export functions. Only the bucket math has
moved. PowerPage imports calculatePowerReport / bucketSocAvg / the
type bag from the new module.
